Fix: Rename Entity to Resource
