+++ src/Resource/Commit.php
@@ -7,7 +7,7 @@
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Localheinz\GitHub\ChangeLog\Entity;
+namespace Localheinz\GitHub\ChangeLog\Resource;
 
 class Commit
 {
